# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mister — il fantacalcio manageriale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: Lista template attivi
      parameters:
        - in: query
          name: active_only
          schema:
            type: boolean
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemplateProfile"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Dettaglio template
      parameters:
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateProfile"
        "404":
          description: Not found
  /superadmin/templates:
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Crea nuovo template (superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateProfile"
  /superadmin/templates/{id}:
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Aggiorna template (superadmin)
      parameters:
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateProfile"
        "404":
          description: Not found
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Elimina template custom (superadmin — solo non-system)
      parameters:
        - $ref: "#/components/parameters/StringId"
      responses:
        "204":
          description: Deleted
        "403":
          description: Template di sistema non eliminabile
        "404":
          description: Not found

  # ─── LEAGUES ─────────────────────────────────────────────────
  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: Lista leghe
      parameters:
        - in: query
          name: admin_user_id
          schema:
            type: string
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
    post:
      operationId: createLeague
      tags: [leagues]
      summary: Crea una lega da template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/{id}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Dettaglio lega
      parameters:
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Not found
    patch:
      operationId: updateLeague
      tags: [leagues]
      summary: Aggiorna lega
      parameters:
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Not found
    delete:
      operationId: deleteLeague
      tags: [leagues]
      summary: Elimina lega
      parameters:
        - $ref: "#/components/parameters/StringId"
      responses:
        "204":
          description: Deleted
  /leagues/{id}/stats:
    get:
      operationId: getLeagueStats
      tags: [leagues]
      summary: Statistiche aggregate della lega
      parameters:
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueStats"
        "404":
          description: Not found

  # ─── FEDERATIONS ─────────────────────────────────────────────
...
  /leagues/{leagueId}/markets:
    get:
      operationId: getLeagueMarkets
      tags: [markets]
      summary: Lista eventi di mercato
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketEvent"
    post:
      operationId: createMarketEvent
      tags: [markets]
      summary: Crea evento di mercato
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketEvent"
  /leagues/{leagueId}/markets/{id}:
    get:
      operationId: getMarketEvent
      tags: [markets]
      summary: Dettaglio evento mercato
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketEvent"
        "404":
          description: Not found
    patch:
      operationId: updateMarketEvent
      tags: [markets]
      summary: Aggiorna evento mercato
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketEventUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketEvent"
        "404":
          description: Not found
    delete:
      operationId: deleteMarketEvent
      tags: [markets]
      summary: Elimina evento mercato
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/StringId"
      responses:
        "204":
          description: Deleted

  # ─── FANTA TEAMS ─────────────────────────────────────────────
...
components:
  parameters:
    StringId:
      in: path
      name: id
      required: true
      schema:
        type: string
    LeagueId:
      in: path
      name: leagueId
      required: true
      schema:
        type: string
  schemas:
...
    FeatureFlags:
      type: object
      description: I 18 feature flag copiati dal template al momento della creazione della lega
      properties:
        multi_season_contracts:
          type: boolean
          description: Contratti pluriennali (>1 stagione)
        max_contract_length:
          type: integer
          description: Durata massima contratto in stagioni (1-10)
        contract_renewal:
          type: boolean
          description: Flusso esplicito di rinnovo contrattuale
        preemption_right:
          type: boolean
          description: Diritto di pareggio in asta per il detentore uscente
        repair_auction_january:
          type: boolean
          description: Asta di riparazione invernale di gennaio
        free_agent_pool:
          type: boolean
          description: Pool svincolati permanente disponibile
        direct_trades:
          type: boolean
          description: Scambi diretti 1-a-1 tra manager
        always_on_markets:
          type: boolean
          description: Mercati e scambi sempre attivi in stagione
        carryover_budget:
          type: boolean
          description: Budget residuo portato alla stagione successiva
        carryover_percentage:
          type: number
          description: Percentuale di budget residuo mantenuta (0-100)
        player_value_dynamic:
          type: boolean
          description: Valore di mercato del giocatore aggiornato dinamicamente
        amortization:
          type: boolean
          description: Ammortamento del prezzo d'acquisto sugli anni di contratto
        release_clauses:
          type: boolean
          description: Clausole rescissorie attive
        clause_default_factor:
          type: number
          description: Moltiplicatore residuo per clausola di default (0.5-1.0)
        rescission_penalty:
          type: boolean
          description: Penale se il manager svincola prima della scadenza
        rescission_recovery_pct:
          type: number
          description: Percentuale crediti recuperata sullo svincolo (0-100)
        no_schema_tactics:
          type: boolean
          description: Formazione senza vincoli di modulo predefinito
        scouting_enabled:
          type: boolean
          description: Sistema di scouting giocatori giovani/emergenti

    # ─── MARKET EVENT SUGGESTION (nel template) ───────────────
    MarketEventSuggestion:
      type: object
      required: [name, type, description]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [auction, trade, release, free_agent]
        mode:
          type: string
          enum: [live, async, blind, token, open]
        window_hint:
          type: string
        description:
          type: string

    # ─── COMPETITION SUGGESTION (nel template) ────────────────
    CompetitionSuggestion:
      type: object
      required: [name, type, description]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [campionato, coppa, battle_royale, sprint_race, formula_uno, punteggio_assoluto]
        description:
          type: string

    # ─── TEMPLATE PROFILE ─────────────────────────────────────
    TemplateProfile:
      type: object
      required: [id, name, tagline, description, complexity_level, estimated_weekly_minutes, icon, feature_flags, suggested_markets, suggested_competitions, is_system, is_active, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        complexity_level:
          type: integer
          description: "1 = base, 2 = intermedio, 3 = avanzato"
        estimated_weekly_minutes:
          type: integer
          description: Tempo stimato di gestione settimanale in minuti
        icon:
          type: string
        feature_flags:
          $ref: "#/components/schemas/FeatureFlags"
        suggested_markets:
          type: array
          items:
            $ref: "#/components/schemas/MarketEventSuggestion"
        suggested_competitions:
          type: array
          items:
            $ref: "#/components/schemas/CompetitionSuggestion"
        author_user_id:
          type: ["string", "null"]
        is_system:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateInput:
      type: object
      required: [name, complexity_level, estimated_weekly_minutes, feature_flags]
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        complexity_level:
          type: integer
        estimated_weekly_minutes:
          type: integer
        icon:
          type: string
        feature_flags:
          $ref: "#/components/schemas/FeatureFlags"
        suggested_markets:
          type: array
          items:
            $ref: "#/components/schemas/MarketEventSuggestion"
        suggested_competitions:
          type: array
          items:
            $ref: "#/components/schemas/CompetitionSuggestion"
        is_active:
          type: boolean
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        complexity_level:
          type: integer
        estimated_weekly_minutes:
          type: integer
        icon:
          type: string
        feature_flags:
          $ref: "#/components/schemas/FeatureFlags"
        suggested_markets:
          type: array
          items:
            $ref: "#/components/schemas/MarketEventSuggestion"
        suggested_competitions:
          type: array
          items:
            $ref: "#/components/schemas/CompetitionSuggestion"
        is_active:
          type: boolean

    # ─── LEAGUE ───────────────────────────────────────────────
    League:
      type: object
      required: [id, name, federation_id, admin_user_id, max_managers, season, visibility, started, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        federation_id:
          type: string
        template_id:
          type: ["string", "null"]
        admin_user_id:
          type: string
        co_admin_user_ids:
          type: array
          items:
            type: string
        max_managers:
          type: integer
        season:
          type: integer
        visibility:
          type: string
          enum: [private, public, unlisted]
        invitation_code:
          type: ["string", "null"]
        lineup_visibility:
          type: string
          enum: [always, after_deadline, hidden_all_season]
        roster_visibility:
          type: string
          enum: [always, after_deadline, hidden_all_season]
        started:
          type: boolean
        notify_email:
          type: boolean
        notify_push:
          type: boolean
        created_at:
          type: string
          format: date-time
    LeagueInput:
      type: object
      required: [name, template_id, admin_user_id, season]
      properties:
        name:
          type: string
        template_id:
          type: string
        admin_user_id:
          type: string
        season:
          type: integer
        max_managers:
          type: integer
        visibility:
          type: string
          enum: [private, public, unlisted]
    LeagueUpdate:
      type: object
      properties:
        name:
          type: string
        max_managers:
          type: integer
        visibility:
          type: string
          enum: [private, public, unlisted]
        lineup_visibility:
          type: string
          enum: [always, after_deadline, hidden_all_season]
        roster_visibility:
          type: string
          enum: [always, after_deadline, hidden_all_season]
        started:
          type: boolean
        notify_email:
          type: boolean
        notify_push:
          type: boolean
    LeagueStats:
      type: object
      required: [league_id, team_count, competition_count, active_market_count, contract_count]
      properties:
        league_id:
          type: string
        team_count:
          type: integer
        competition_count:
          type: integer
        active_market_count:
          type: integer
        contract_count:
          type: integer

    # ─── FEDERATION RULES ─────────────────────────────────────
...
    MarketEvent:
      type: object
      required: [id, league_id, name, type, status, starts_at, ends_at, config, created_at]
      properties:
        id:
          type: string
        league_id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [auction, trade, release, free_agent]
          description: "auction = asta (sub-mode in config), trade = scambi, release = svincoli, free_agent = pool svincolati"
        status:
          type: string
          enum: [scheduled, active, completed, cancelled]
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        config:
          type: object
          description: "Configurazione specifica per tipo: AuctionRules / TradeRules / ReleaseRules / FreeAgentRules + labelColor"
        created_at:
          type: string
          format: date-time
    MarketEventInput:
      type: object
      required: [name, type, window_starts_at, window_ends_at]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [auction, trade, release, free_agent]
        window_starts_at:
          type: string
          format: date-time
        window_ends_at:
          type: string
          format: date-time
        settings:
          type: object
        label_color:
          type: string
    MarketEventUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [scheduled, active, completed, cancelled]
        window_starts_at:
          type: string
          format: date-time
        window_ends_at:
          type: string
          format: date-time
        settings:
          type: object

    # ─── FANTA TEAM ───────────────────────────────────────────
...