# lib/api-zod/src/generated/types/marketEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mister — il fantacalcio manageriale
 * OpenAPI spec version: 0.1.0
 */
import type { MarketEventConfig } from './marketEventConfig';
import type { MarketEventStatus } from './marketEventStatus';
import type { MarketEventType } from './marketEventType';

export interface MarketEvent {
  id: string;
  league_id: string;
  name: string;
  description?: string;
  /** auction = asta (sub-mode in config), trade = scambi, release = svincoli, free_agent = pool svincolati */
  type: MarketEventType;
  status: MarketEventStatus;
  starts_at: Date;
  ends_at: Date;
  /** Configurazione specifica per tipo: AuctionRules / TradeRules / ReleaseRules / FreeAgentRules + labelColor */
  config: MarketEventConfig;
  created_at: Date;
}