# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mister — il fantacalcio manageriale
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitionSuggestion } from './competitionSuggestion';
import type { FeatureFlags } from './featureFlags';
import type { MarketEventSuggestion } from './marketEventSuggestion';

export interface TemplateInput {
  name: string;
  tagline?: string;
  description?: string;
  complexity_level: number;
  estimated_weekly_minutes: number;
  icon?: string;
  feature_flags: FeatureFlags;
  suggested_markets?: MarketEventSuggestion[];
  suggested_competitions?: CompetitionSuggestion[];
  is_active?: boolean;
}